require iconv for SUPPORT_MBCS tidy up a bit in the non-MBCS case
git-svn-id: https://svn.r-project.org/R/trunk@35947 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/localecharset.c
@@ -668,7 +668,9 @@ char *locale2charset(const char *locale)
 
 	/* let's hope it is a ll_* name */
 	if (0 == strcmp(enc, "euc")) {
-	    if(isalpha(enc[0]) && isalpha(enc[1]) && (enc[2] == '_')) {
+	    /* This is OK as encoding names are ASCII */
+	    if(isalpha((int)enc[0]) && isalpha((int)enc[1]) 
+	       && (enc[2] == '_')) {
 		if (0 == strncmp("ja", la_loc, 2)) return "EUC-JP";
 		if (0 == strncmp("ko", la_loc, 2)) return "EUC-KR";
 		if (0 == strncmp("zh", la_loc, 2)) return "GB2312";
